z80asm: consistent comment style

+++ z80asm/z80anum.c
@@ -132,11 +132,8 @@ void init_ctype(void)
 }
 
 /*
- *	binary search in sorted table oprtab
- *
- *	Input: pointer to string with operator
- *
- *	Output: symbol for operator, T_UNDSYM if operator not found
+ *	do binary search for operator s in sorted table oprtab
+ *	returns symbol for operator or T_UNDSYM if not found
  */
 BYTE search_opr(char *s)
 {
@@ -159,8 +156,7 @@ BYTE search_opr(char *s)
 
 /*
  *	get next token
- *
- *	updates tok_type, tok_val, tok_sym and scan_pos.
+ *	updates tok_type, tok_val, tok_sym and scan_pos
  *	returns E_NOERR on success
  */
 int get_token(void)
@@ -239,7 +235,7 @@ int get_token(void)
 			tok_type = T_VAL;
 			tok_val = pc;
 		} else {				/* symbol / word opr */
-			if (p2 - p1 > symlen)		/* trim for lookup */
+			if ((unsigned) (p2 - p1) > symlen) /* trim to symlen */
 				*(p1 + symlen) = '\0';
 			if ((sp = get_sym(tok_sym)) != NULL) {	/* a symbol */
 				tok_type = T_VAL;
@@ -619,7 +615,7 @@ WORD eval(char *s)
 
 /*
  *	check value for range -256 <= value <= 255
- *	Output: value if in range, otherwise 0 and error message
+ *	returns value if in range, otherwise 0 and error message
  */
 BYTE chk_byte(WORD n)
 {
@@ -633,7 +629,7 @@ BYTE chk_byte(WORD n)
 
 /*
  *	check value for range -128 <= value <= 127
- *	Output: value if in range, otherwise 0 and error message
+ *	returns value if in range, otherwise 0 and error message
  */
 BYTE chk_sbyte(WORD n)
 {
